requirements.txt
anthropic==0.29.0
html2text==2024.2.26
pdf2image==1.17.0
Pillow==9.4.0
Pillow==10.3.0
playwright==1.33.0
python-dotenv==1.0.1
pytrends==4.9.2
pytz==2023.3
Requests==2.32.3
retrying==1.3.4
